pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "Page Analyzer Flask App"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2>=2.9.10",
    "python-dotenv>=1.1.1",
    "validators>=0.35.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["page_analyzer"]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-cov>=6.0.0",
    "ruff>=0.9.10",
]
